osd/: factor OSD::init_op_flags into seperate class
We'll want to reuse this in crimson, extract the logic
for setting flags from MOSDOp into osd_types.h.
