+++ src/osd/OSDCap.cc
@@ -249,7 +249,7 @@ bool OSDCapGrant::is_capable(
   const string& object,
   bool op_may_read,
   bool op_may_write,
-  const std::vector<OpRequest::ClassInfo>& classes,
+  const std::vector<OpInfo::ClassInfo>& classes,
   const entity_addr_t& addr,
   std::vector<bool>* class_allowed) const
 {
@@ -377,7 +377,7 @@ bool OSDCap::is_capable(const string& pool_name, const string& ns,
 			const OSDCapPoolTag::app_map_t& application_metadata,
 			const string& object,
                         bool op_may_read, bool op_may_write,
-			const std::vector<OpRequest::ClassInfo>& classes,
+			const std::vector<OpInfo::ClassInfo>& classes,
 			const entity_addr_t& addr) const
 {
   std::vector<bool> class_allowed(classes.size(), false);
